requirements.txt
ollama
requests
telebot
python-dotenv